+++ src/server.rs
@@ -11,7 +11,7 @@ use core::{ops::Range, time::Duration};
 use bevy::{
     ecs::{
         archetype::Archetypes,
-        component::{StorageType, Tick},
+        component::Tick,
         entity::{Entities, EntityHashMap},
         intern::Interned,
         schedule::ScheduleLabel,
@@ -622,21 +622,6 @@ fn collect_changes(
                 mutations.start_entity();
             }
 
-            // SAFETY: all replicated archetypes have marker component with table storage.
-            let (_, marker_ticks) = unsafe {
-                world.get_component_unchecked(
-                    entity,
-                    archetype.table_id(),
-                    StorageType::Table,
-                    world.marker_id(),
-                )
-            };
-            // If the marker was added in this tick, the entity just started replicating.
-            // It could be a newly spawned entity or an old entity with just-enabled replication,
-            // so we need to include even old components that were registered for replication.
-            let marker_added =
-                marker_ticks.is_added(change_tick.last_run(), change_tick.this_run());
-
             for &(component_rule, storage) in &replicated_archetype.components {
                 let (component_id, component_fns, rule_fns) = registry.get(component_rule.fns_id);
 
@@ -664,8 +649,6 @@ fn collect_changes(
                     }
 
                     if let Some((last_system_tick, last_server_tick)) = entity_cache.mutation_tick
-                        && !marker_added
-                        && entity_cache.visibility != Visibility::Gained
                         && !ticks.is_added(change_tick.last_run(), change_tick.this_run())
                     {
                         let tick_diff = server_tick - last_server_tick;
@@ -732,8 +715,7 @@ fn collect_changes(
                     continue;
                 }
 
-                let new_entity = marker_added || entity_cache.visibility == Visibility::Gained;
-                if new_entity
+                if entity_cache.mutation_tick.is_none()
                     || updates.changed_entity_added()
                     || removal_buffer.contains_key(&entity.id())
                 {
@@ -749,7 +731,7 @@ fn collect_changes(
                     ticks.set_mutation_tick(entity.id(), change_tick.this_run(), server_tick);
                 }
 
-                if new_entity && !updates.changed_entity_added() {
+                if entity_cache.mutation_tick.is_none() && !updates.changed_entity_added() {
                     trace!(
                         "writing empty `{}` for client `{client_entity}`",
                         entity.id()
@@ -778,9 +760,7 @@ fn should_send_mapping(
         return false;
     }
 
-    visibility_state == Visibility::Gained
-        || signature.is_added()
-        || ticks.mutation_tick(entity).is_none()
+    signature.is_added() || ticks.mutation_tick(entity).is_none()
 }
 
 /// Writes a mapping or re-uses previously written range if exists.