+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `Debug` derive for `FromClient`.
 
+### Fixed
+
+- Replication of entities without components before connection.
+
 ## [0.35.0] - 2025-09-26
 
 ### Added